refactor: split into user steps
